[refractor/postgres load][xl]: direct postgres copy loader implemented and code refractored
